+++ src/main/git/gitTypes.ts
@@ -1,0 +1,16 @@
+export interface Repository {
+    name: string,
+    author: string,
+    url: string
+}
+
+export enum Url {
+    REDIRECT_URI = "http://localhost:3000/main_window",
+    AUTHORIZE_URL = "https://github.com/login/oauth/authorize",
+    ACCESS_TOKEN_URL = "https://github.com/login/oauth/access_token"
+}
+
+export enum EndPointParameters {
+    CLIENT_ID = "client_id",
+    REDIRECT_URI ="redirect_uri"
+}
